Виправлення типу полів

+++ database/migrations/2025_05_03_165459_create_role_limits_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('role_limits', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->string('name');
-            $table->integer('value');
+            $table->unsignedBigInteger('value');
             $table->timestamps();
 
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
