+++ database/migrations/2025_05_03_165528_create_site_user_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('site_user', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('site_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('site_id');
             $table->enum('role', ['owner', 'user']);
             $table->timestamps();
